+++ croud/subscriptions/commands.py
@@ -34,7 +34,11 @@ def subscriptions_get(args: Namespace) -> None:
 
 def subscriptions_list(args: Namespace) -> None:
     client = Client.from_args(args)
-    data, errors = client.get("/api/v2/subscriptions/")
+    if args.org_id:
+        url = f"/api/v2/organizations/{args.org_id}/subscriptions/"
+    else:
+        url = "/api/v2/subscriptions/"
+    data, errors = client.get(url)
     print_response(
         data=data,
         errors=errors,